ConfigurationLock

+++ Authentication/Application/OTPService.cs
@@ -2,17 +2,17 @@
 namespace Authentication.Application;
 public class OtpService
 {
-    private readonly Dictionary<int, string> _otpStore = new();
+    private readonly Dictionary<long, string> _otpStore = new();
     private readonly Random _random = new();
 
-    public string GenerateOtp(int userId)
+    public string GenerateOtp(long userId)
     {
         var otp = _random.Next(100000, 999999).ToString();
         _otpStore[userId] = otp;
         return otp;
     }
 
-    public bool ValidateOtp(int userId, string otpCode)
+    public bool ValidateOtp(long userId, string otpCode)
     {
         return _otpStore.ContainsKey(userId) && _otpStore[userId] == otpCode;
     }
